% Generated by roxygen2 (4.0.2): do not edit by hand
\name{google_trends}
\alias{google_trends}
\title{Get data from Google Trends}
\usage{
google_trends(query, country = "all", region = "all", date = "all",
  category = "all", sub_category = "all")
}
\arguments{
\item{query}{The keyword to be searched. It will be auto escaped.}

\item{country}{Country name, i.e: United Kingdom, Italy, Germany, etc. Deafult option is the worldwide search.}

\item{region}{Region name, i.e: England, Scotland, etc. Deafult option is all. It is possible to use only the region name without specifying the related country.}

\item{date}{The time span. Default is all. Other options are: 'last 7 days', 'last 30 days', 'last 90 days', 'last year' or the year number, i.e '2014'}

\item{category}{The parent category of the keyword. Default is all. The options are listed in the data GT_Options$Category dataframe}

\item{sub_category}{The child category of the keywrod. Default is all.}
}
\value{
The output is a list. This list contains the trend, the geographical segmentation (countries and cities), the suggested categories for the keyword and the related searches
}
\description{
This function scrapes Google Trends and returns the results nicely formatted into a R object
}
\note{
Please note that (except for the query argument) all the other arguments are case INSENSITIVE
}
\examples{
#Search the keyword 'test me' in England for last 30 days
results = google_trends(query = "test me", region = "England", date = "last 30 days")

#Access all available arguments
data(Google_Trends_Data)
#i.e: Check available categories
unique(GT_Options$Category$name)
}

