% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{japanGDPData}
\alias{japanGDPData}
\title{Loads Japan Gross Domestic Product (GDP) data set}
\format{A data frame with 59 rows and 2 variables:
\describe{
  \item{year}{The year.}
  \item{value}{The GDP value.}
}}
\usage{
japanGDPData
}
\description{
A subset of the World Banks Gross Domestic Product (GDP) data set for 1960 to 2018
}
\keyword{datasets}
