% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodLLMvard.R
\name{likelihoodLLMcard}
\alias{likelihoodLLMcard}
\title{This the function computes the likelihood for local linear model - steady state model with a time varying drift (second component of alpha)}
\usage{
likelihoodLLMcard(x, y)
}
\arguments{
\item{x}{The values as a xtor}

\item{y}{The timeseries}
}
\description{
This the function computes the likelihood for local linear model - steady state model with a time varying drift (second component of alpha)
}
\author{
Sonia Mazzi
}
