% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naver_trends.R
\name{naver_trends}
\alias{naver_trends}
\title{Get data from Naver Trends}
\usage{
naver_trends(query)
}
\arguments{
\item{query}{The keyword to be searched. It will be auto escaped. This can be a vector with up to 5 queries}
}
\value{
The output is a dataframe containing Date and Value variables for all the queries
}
\description{
This function scrapes Naver Trends and returns the results nicely formatted into a R object
}
\examples{
#Search the keyword 'tokyo and japan'
results = naver_trends(c("tokyo", "japan"))

#Visualize results
ggplot(results, aes(x = Date, y = Value, col = Query)) + geom_line()

}

