% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{randomData}
\alias{randomData}
\title{Loads example data set}
\format{A data frame with 41 rows and 2 variables:
\describe{
  \item{date}{The date.}
  \item{value}{The random value.}
}}
\usage{
randomData
}
\description{
A random data set
}
\keyword{datasets}
