% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfilter.R
\name{smfilt}
\alias{smfilt}
\title{This is the smoothing filter. Relies on DFK calculated first.}
\usage{
smfilt(dkf.out)
}
\arguments{
\item{dkf.out}{The values as a vector}
}
\description{
This is the smoothing filter. Relies on DFK calculated first.
}
\author{
Sonia Mazzi
}
