% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_360_trends.R
\name{so_360_trends}
\alias{so_360_trends}
\title{Get data from SO 360 (Chinese Search Engine)}
\usage{
so_360_trends(query)
}
\arguments{
\item{query}{The keyword to be searched.}
}
\value{
The output is a dataframe containing Date and Value variables for all the queries
}
\description{
This function scrapes 360 Trends and returns the results nicely formatted into a R object
}
\examples{
#Search the keyword 'tokyo and japan'
results = so_360_trends("洛杉矶酒店")

#Visualize results
ggplot(results, aes(x = Date, y = Value, col = Query)) + geom_line()

}

