% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendr.R
\name{trendr}
\alias{trendr}
\title{Calculates the trend and first derivative using a local linear model}
\usage{
trendr(df = randomData, value.colname = "count",
  time.colname = "time", output.dir, output.file = "out.csv",
  output.plot = F)
}
\arguments{
\item{df}{The dataframe that comprises a time series}

\item{value.colname}{The column name comprising the values}

\item{time.colname}{The column name comprising the timeseries values}

\item{output.dir}{The directory for the output file}

\item{output.file}{The file name for the output file}

\item{output.plot}{Specify whether you want the output to be plotted using R's basic plotting}
}
\description{
Calculates the trend and first derivative using a local linear model. Also uses a Kalman
Filter.
}
\examples{
  #trendr(
    #df = randomData,
    #value.colname = 'count',
    #time.colname = 'time',
    #output.dir = 'C:\\Users\\User\\Documents\\',
    #output.file = 'out.csv',
    #output.plot = T)
}
\author{
Michael Hodge
}
