% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendr_beta.R
\name{trendr_beta}
\alias{trendr_beta}
\title{Calculates the first derivative using a local linear model}
\usage{
trendr_beta(y)
}
\arguments{
\item{y}{the values}
}
\description{
Calculates the trend and first derivative using a local linear model. Also uses a Kalman
Filter.
}
\author{
Michael Hodge
}
