% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendr_mu.R
\name{trendr_mu}
\alias{trendr_mu}
\title{Calculates the trend and first derivative using a local linear model}
\usage{
trendr_mu(y)
}
\arguments{
\item{y}{the values}
}
\description{
Calculates the trend and first derivative using a local linear model. Also uses a Kalman
Filter.
}
\author{
Michael Hodge
}
