% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukRainData}
\alias{ukRainData}
\title{Loads UK rainfall data set}
\format{A data frame with 33 rows and 4 variables:
\describe{
  \item{year}{The year.}
  \item{rain_year}{The average monthly rainfall in mm.}
  \item{rain_winter}{The average monthly rainfall during winter months in mm.}
  \item{rain_summer}{The average monthly rainfall during summer months in mm.}
}}
\usage{
ukRainData
}
\description{
A subset of DEFRA's UK rainfall data for 1980 to 2012
}
\keyword{datasets}
