% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkfLLMvard.R
\name{dfkLLMvard}
\alias{dfkLLMvard}
\title{Calculates the Dynamic Kalman Filter (DFK).}
\usage{
dfkLLMvard(x, y)
}
\arguments{
\item{x}{The values as a xtor}

\item{y}{The timeseries}
}
\description{
The state space form is
y(x)  =  Z(x)*alpha(x) + G*u(x)
alpha(x+1) = TT*alpha(x) + H*u(x)
the DKF is initialized with A0 and P0 (Q0=0).
}
\details{
sigma.eps^2 is concentrated out
}
\author{
Sonia Mazzi
}
