% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTrends}
\alias{getTrends}
\title{Get Trend data from Google}
\usage{
getTrends(query, country = "all", region = "none", date = "all")
}
\arguments{
\item{query}{The keyword to be searched. It will be auto escaped.}

\item{country}{Country name, i.e: United Kingdom, Italy, Germany, etc. Case INSENSITIVE. Deafult option is the worldwide research.}

\item{region}{Region name, i.e: England, Scotland, etc. Case INSENSITIVE. Deafult option is none.}

\item{date}{The time span. Default is all. Other options are: 'last 7 days', 'last 30 days', 'last 90 days', 'last year' or the year number, i.e '2014'}
}
\value{
data frame
}
\description{
This function scrapes Google Trends and returns the results nicely formatted in a R dataframe
}
\examples{
getTrends(query = "test me", country = "United Kingdom", region = "England", date = "last 30 days")
}

