% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfilter.R
\name{smfilt}
\alias{smfilt}
\title{This is the smoothing filter. The DFK must be run first.}
\usage{
smfilt(dkf.out)
}
\arguments{
\item{dkf.out}{A list (output of the DKF)}
}
\description{
This is the smoothing filter. The DFK must be run first.
}
\author{
Sonia Mazzi
}
