% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendr.R
\name{trendr}
\alias{trendr}
\title{Extracts the trend and first derivative using a local linear model in state-space form and the Diffuse Kalman
Filter.}
\usage{
trendr(df = randomData, value.colname = "value",
  time.colname = "date", output.save = FALSE, output.dir = "./",
  output.file = "out.csv", output.plot = FALSE)
}
\arguments{
\item{df}{The data frame with a time series}

\item{value.colname}{The name of the column with the values of the series}

\item{time.colname}{The name of the column with the time values}

\item{output.save}{Whether the results should save locally or now (default is FALSE)}

\item{output.dir}{The directory for the output file, by default your working directory}

\item{output.file}{The file name for the output file, by default out.csv}

\item{output.plot}{Specify whether you want the output to be plotted using R's basic plotting}
}
\description{
Extracts the trend and first derivative using a local linear model in state-space form and the Diffuse Kalman
Filter.
}
\examples{
  #trendr(
    #df = randomData,
    #value.colname = 'count',
    #time.colname = 'time',
    #output.dir = 'C:\\Users\\User\\Documents\\',
    #output.file = 'out.csv',
    #output.plot = T)
}
\author{
Sonia Mazzi and Michael Hodge
}
